Fix get_config_uint64() to read a 64-bit value not a 32-bit one.

+++ lib/config/config.h
@@ -31,7 +31,7 @@ enum {
 struct config_value {
 	int type;
 	union {
-		int i;
+		int64_t i;
 		float r;
 		char *str;
 	} v;
